fix(extension): preserve bundled paths on POSIX
