+++ docs/features/active/2026-04-05-ci-failing-error-128/audit-2026-04-05T20-20/code-review.2026-04-05T20-20.md
@@ -1,0 +1,83 @@
+# Code Review: 2026-04-05-ci-failing-error-128
+
+## Executive Summary
+
+The reviewed working-tree change fixes cross-platform bundled-path resolution by updating `resolveBundledScriptPath` in `extensions/drm-copilot/src/command-runtime.ts` to preserve Windows drive-prefixed roots such as `C:/extension` when the host process uses POSIX path semantics. The change also adds four focused regressions in `extensions/drm-copilot/test/extension.test.ts` and `extensions/drm-copilot/test/repo-automation-service.test.ts` for `helloPython`, `helloPowerShell`, `collectCommitContext`, and `newPotentialEntry`.
+
+The refreshed PR-context artifacts were anchored to `development`, and fresh review checks passed: non-mutating Prettier check, ESLint, TSC, targeted regression tests, full Jest unit suite, and Jest coverage. Functional behavior is correct and the three acceptance criteria are satisfied.
+
+### Top Risks
+
+1. `extensions/drm-copilot/test/extension.test.ts` is now 918 lines, which violates the repository’s 500-line file limit for touched test files.
+2. The new POSIX-path fresh-module helpers are duplicated across both touched test files, which increases maintenance drift risk.
+3. The committed branch range still reflects only the initial docs/baseline commit; the functional fix currently exists in the working tree and must be preserved carefully during remediation.
+
+### PR Readiness Recommendation
+
+**No-Go / Needs revision** until the touched oversized test file is brought back into policy compliance.
+
+## Scope Anchor
+
+**Feature Folder:** `docs/features/active/2026-04-05-ci-failing-error-128/`  
+**Selection Rule:** Used the explicitly provided feature folder and confirmed the same folder in the refreshed `artifacts/pr_context.summary.txt` additional-context list.  
+**Base Branch:** `development`
+
+## Findings
+
+| Severity | File | Location | Finding | Recommendation | Rationale | Evidence |
+|---|---|---|---|---|---|---|
+| Blocker | `extensions/drm-copilot/test/extension.test.ts` | file-wide; new helper block starts near line 220; new regressions near lines 489 and 526 | The touched test file is 918 lines after this change, exceeding the repo-wide 500-line limit that applies to test code. | Split or reorganize the touched test coverage so every touched file is under 500 lines. A shared helper for the POSIX fresh-module path setup is the most likely first extraction point. Preserve the new regression scenario names and expectations. | This is a direct policy violation on a touched file and the change worsened the pre-existing condition by adding 120 lines. | Fresh line count check: `extensions/drm-copilot/test/extension.test.ts: 918`; working-tree diff stat: `+120/-0`; general policy caps test files at 500 lines. |
+| Minor | `extensions/drm-copilot/test/extension.test.ts`, `extensions/drm-copilot/test/repo-automation-service.test.ts` | helper blocks near lines 220 and 38 respectively | The fresh-module POSIX-path mocking helpers are duplicated across both test files. | Consolidate the shared helper logic during the file-size remediation, either by extracting a shared test utility or by restructuring the test layout so the helper is defined once. | Duplication increases future maintenance cost and was a major contributor to the line-count growth in the touched extension command test file. | Matching helper names and logic appear in both files: `prepareFreshModulesWithPosixPathResolve`, `setFreshExecutablePresence`, and child-process helper accessors. |
+
+## Verified Strengths
+
+- `extensions/drm-copilot/src/command-runtime.ts` keeps the fix local to the shared path-resolution helper instead of scattering platform checks across call sites.
+- The new regression scenarios precisely model the CI failure mode by forcing POSIX `path.resolve` semantics with Windows-style mocked `extensionUri.fsPath` values.
+- Fresh verification passed:
+  - Prettier check
+  - ESLint
+  - TSC
+  - targeted regression run (`26 passed, 25 skipped, 51 total`)
+  - full unit suite (`140 passed, 140 total`)
+  - coverage (`88.30%` lines overall; `93.46%` lines for `src/command-runtime.ts`)
+
+## TypeScript Review
+
+### Design and Typing
+
+- No new `any` usage was introduced in the reviewed files.
+- The production fix remains a typed, deterministic branch inside `resolveBundledScriptPath`.
+- No public API was broken; `resolveBundledScriptPath` remains exported.
+
+### Suppressions
+
+- No new broad suppressions were introduced.
+- Existing single-line ESLint suppressions in tests remain narrowly scoped and justified.
+
+### Error Handling and Reliability
+
+- The change does not weaken subprocess safety or shell handling.
+- The new logic is constant-time string normalization and regex matching.
+- Existing runtime failure-path tests continue to pass in the full suite.
+
+## Test Quality Review
+
+- **Deterministic:** The new tests use mocked modules and fixed path fixtures.
+- **Isolated:** Each regression targets one command/service scenario.
+- **Fast:** The targeted regression run completed in well under one second.
+- **Diagnostics:** Scenario names clearly identify the cross-platform failure mode.
+- **Coverage:** The changed production file exceeds the 90% threshold in the collected coverage summary.
+
+## Security and Correctness Checks
+
+- No secrets or credentials were added.
+- No unsafe `shell: true` subprocess behavior was introduced.
+- The fix preserves bundled extension resource targeting and avoids checkout-prefixed hybrid paths on POSIX hosts.
+
+## Research Log
+
+No external research was required for this review.
+
+## Recommendation
+
+The bug fix is behaviorally correct, but the branch should not be treated as PR-ready until the touched oversized test file is brought back into compliance with the repository’s file-size policy and the verification artifacts are refreshed afterward.
