+++ docs/features/active/2026-04-05-ci-failing-error-128/audit-2026-04-05T20-20/feature-audit.2026-04-05T20-20.md
@@ -1,0 +1,53 @@
+# Feature Audit: 2026-04-05-ci-failing-error-128
+
+## Scope and Baseline
+
+- **Base Branch:** `development`
+- **Head Branch:** `bug/ci-failing-error-128`
+- **Feature Folder:** `docs/features/active/2026-04-05-ci-failing-error-128/`
+- **Feature Folder Selection Rule:** Used the user-provided active feature folder and confirmed the same folder in refreshed PR-context additional-context entries.
+- **Evidence Sources:**
+  - Primary: `artifacts/pr_context.summary.txt` (refreshed during this review)
+  - Secondary: `artifacts/pr_context.appendix.txt` (working-tree diff and hunk evidence)
+  - Feature evidence: `docs/features/active/2026-04-05-ci-failing-error-128/evidence/**`
+- **Work Mode:** `minor-audit`
+- **Authoritative AC Source:** `docs/features/active/2026-04-05-ci-failing-error-128/issue.md#acceptance-criteria`
+- **Non-required Docs Check:** `spec.md` absent; `user-story.md` absent. This satisfies the reduced-audit requirement.
+
+## Acceptance Criteria Inventory
+
+Authoritative acceptance criteria extracted from `issue.md`:
+
+1. Windows-style absolute extension roots such as `C:/extension` remain absolute when bundled script paths are resolved on POSIX hosts.
+2. Bundled command and repo-automation script invocations continue to target extension resources rather than workspace-relative copies.
+3. Regression coverage verifies the Windows-style mocked `fsPath` scenario and the extension Jest suite no longer fails with checkout-prefixed hybrid paths.
+
+## Acceptance Criteria Evaluation
+
+| Criterion | Status | Evidence | Verification Command(s) | Notes |
+|---|---|---|---|---|
+| Windows-style absolute extension roots such as `C:/extension` remain absolute when bundled script paths are resolved on POSIX hosts. | PASS | `extensions/drm-copilot/src/command-runtime.ts` adds a drive-root guard in `resolveBundledScriptPath`; `p1-t8.green-jest.2026-04-05T20-12.md`; `p2-t4.test-unit.2026-04-05T20-14.md` | `Push-Location extensions/drm-copilot; npm run test:unit -- --runTestsByPath test/extension.test.ts test/repo-automation-service.test.ts -t "helloPython|helloPowerShell|collectCommitContext|newPotentialEntry"; Pop-Location` | Fresh targeted run passed all four Windows-root scenarios. |
+| Bundled command and repo-automation script invocations continue to target extension resources rather than workspace-relative copies. | PASS | `helloPython`, `helloPowerShell`, `collectCommitContext`, and `newPotentialEntry` regressions all assert `C:/extension/resources/templates/...`; `p1-t8.green-jest.2026-04-05T20-12.md`; fresh targeted run output | `Push-Location extensions/drm-copilot; npm run test:unit -- --runTestsByPath test/extension.test.ts test/repo-automation-service.test.ts -t "helloPython|helloPowerShell|collectCommitContext|newPotentialEntry"; Pop-Location` | The fresh targeted run and on-disk green-run artifact both confirm bundled-resource targeting. |
+| Regression coverage verifies the Windows-style mocked `fsPath` scenario and the extension Jest suite no longer fails with checkout-prefixed hybrid paths. | PASS | `p1-t6.red-jest.2026-04-05T20-11.md`; `p1-t8.green-jest.2026-04-05T20-12.md`; `p2-t4.test-unit.2026-04-05T20-14.md`; fresh full Jest run (`140/140`) | `Push-Location extensions/drm-copilot; npm run test:unit; Pop-Location` | The feature has fail-before and pass-after evidence plus a fresh passing full-suite verification. |
+
+## Summary
+
+**Overall feature readiness:** PASS
+
+The implemented behavior satisfies all three acceptance criteria defined in the explicit `## Acceptance Criteria` section of `issue.md`. The required reduced-audit structure is intact: `spec.md` and `user-story.md` are absent, Phase 0 baseline evidence exists, the plan checklist is backed by on-disk artifacts and current code/test state, and the small-path production scope remains limited to `extensions/drm-copilot/src/command-runtime.ts`.
+
+**Top gaps preventing PR readiness:** None at the feature-behavior level. A separate policy-compliance issue remains, documented in `policy-audit.2026-04-05T20-20.md` and `code-review.2026-04-05T20-20.md`.
+
+**Recommended follow-up verification:** None required for acceptance behavior. Policy remediation is still required before PR readiness.
+
+## Acceptance Criteria Status
+
+- Source: `docs/features/active/2026-04-05-ci-failing-error-128/issue.md`
+- Total AC items: 3
+- Checked off (delivered): 3
+- Remaining (unchecked): 0
+- Items remaining: None
+
+## Acceptance Criteria Check-Off
+
+No additional checkbox edits were required during this review because `issue.md` already had all three authoritative acceptance-criteria items marked `[x]`, and the review evidence supports those checked states.
