+++ docs/features/active/2026-04-05-ci-failing-error-128/audit-2026-04-05T20-42/code-review.2026-04-05T20-42.md
@@ -1,0 +1,27 @@
+# Code Review: 2026-04-05-ci-failing-error-128 (Remediation Refresh)
+
+Review Timestamp: 2026-04-05T20:42:28.8607822-04:00
+
+## Executive Summary
+
+The remediation reorganized the touched extension test surface without altering the existing production fix. Shared POSIX fresh-module helpers were extracted into dedicated test helper modules, `extension.test.ts` was reduced to the core hello-command coverage plus the two preserved Windows-root POSIX regressions, and the remaining command coverage moved into `extension.workflow-commands.test.ts`.
+
+## Preserved Regression Scenarios
+
+- `helloPython preserves C:/extension on POSIX hosts`
+- `helloPowerShell preserves C:/extension on POSIX hosts`
+- `collectCommitContext preserves C:/extension on POSIX hosts`
+- `newPotentialEntry preserves C:/extension on POSIX hosts`
+
+## Final Verification
+
+- Prettier check: PASS
+- ESLint: PASS
+- TypeScript type-check: PASS
+- Targeted regressions: PASS
+- Full Jest coverage run: PASS
+- Coverage headline: 89.36% lines, 89.36% statements, 88.69% functions, 81.61% branches
+
+## Review Conclusion
+
+No touched test/helper file exceeds 500 lines, the existing production fix in `extensions/drm-copilot/src/command-runtime.ts` was not modified during remediation, and the refreshed evidence supports PR readiness if no unrelated issues remain.
