+++ docs/features/active/2026-04-05-ci-failing-error-128/audit-2026-04-05T20-42/policy-audit.2026-04-05T20-42.md
@@ -1,0 +1,30 @@
+# Policy Compliance Audit: 2026-04-05-ci-failing-error-128 (Remediation Refresh)
+
+Audit Timestamp: 2026-04-05T20:42:28.8607822-04:00
+Work Mode: minor-audit
+Acceptance Criteria Source: docs/features/active/2026-04-05-ci-failing-error-128/issue.md#acceptance-criteria
+
+## Summary
+
+The remediation refresh resolves the blocking 500-line policy violation without changing production code. The current fix in `extensions/drm-copilot/src/command-runtime.ts` was left untouched, the four named Windows-root POSIX regression scenarios remain present exactly as named, and the touched extension test/helper files are now all within the repository 500-line limit.
+
+## Final Touched Test/Helper File Sizes
+
+- `extensions/drm-copilot/test/extension.test.ts` — 251 lines
+- `extensions/drm-copilot/test/extension.workflow-commands.test.ts` — 393 lines
+- `extensions/drm-copilot/test/extension-test-harness.ts` — 244 lines
+- `extensions/drm-copilot/test/runtime-test-helpers.ts` — 135 lines
+- `extensions/drm-copilot/test/repo-automation-service.test.ts` — 234 lines
+
+## Verification Evidence
+
+- `evidence/final-qa/final-prettier-check.md`
+- `evidence/final-qa/final-eslint.md`
+- `evidence/final-qa/final-tsc.md`
+- `evidence/final-qa/final-targeted-regressions.md`
+- `evidence/final-qa/final-jest-coverage.md`
+- `evidence/final-qa/final-scope-and-line-counts.md`
+
+## Verdict
+
+PASS. No touched test/helper file remains over 500 lines, no remediation-time production edits were introduced, and the branch is PR-ready if no unrelated issues remain.
