+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-eslint.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:41:44.1449726-04:00
+Command: Push-Location extensions/drm-copilot; npm run lint; Pop-Location
+EXIT_CODE: 0
+Output Summary: ESLint passed with no blocking diagnostics during the clean consecutive QA pass.
