+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-jest-coverage.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:42:14.7874757-04:00
+Command: Push-Location extensions/drm-copilot; npx jest --coverage --coverageReporters=text-summary --runInBand; Pop-Location
+EXIT_CODE: 0
+Output Summary: Full extension Jest coverage passed during the clean consecutive QA pass with 11 passed suites and 140 passed tests. Coverage headline: lines 89.36%, statements 89.36%, functions 88.69%, branches 81.61%. Because remediation touched only test/helper files and left extensions/drm-copilot/src/command-runtime.ts unchanged, the prior measured command-runtime.ts coverage remains the relevant production-file reference at 93.46% lines and 92.30% functions.
