+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-prettier-check.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:41:30.4171791-04:00
+Command: Push-Location extensions/drm-copilot; npx prettier --check 'src/**/*.ts' 'test/**/*.ts' '*.json' '*.cjs'; Pop-Location
+EXIT_CODE: 0
+Output Summary: Final Prettier check passed during the clean consecutive QA pass. All matched extension TypeScript, test, JSON, and CJS files use Prettier code style.
