+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-scope-and-line-counts.md
@@ -1,0 +1,25 @@
+Timestamp: 2026-04-05T20:42:28.8607822-04:00
+Final Touched Test/Helper Files:
+- extensions/drm-copilot/test/extension.test.ts
+- extensions/drm-copilot/test/extension.workflow-commands.test.ts
+- extensions/drm-copilot/test/extension-test-harness.ts
+- extensions/drm-copilot/test/runtime-test-helpers.ts
+- extensions/drm-copilot/test/repo-automation-service.test.ts
+Final Line Counts:
+- extensions/drm-copilot/test/extension.test.ts = 251
+- extensions/drm-copilot/test/extension.workflow-commands.test.ts = 393
+- extensions/drm-copilot/test/extension-test-harness.ts = 244
+- extensions/drm-copilot/test/runtime-test-helpers.ts = 135
+- extensions/drm-copilot/test/repo-automation-service.test.ts = 234
+Final Scope Checks:
+- Every touched test/helper file is <= 500 lines.
+- No production files were changed during remediation.
+- extensions/drm-copilot/src/command-runtime.ts remains unchanged from the current fix.
+- The exact Windows-root POSIX regression scenario names remain present:
+  - helloPython preserves C:/extension on POSIX hosts
+  - helloPowerShell preserves C:/extension on POSIX hosts
+  - collectCommitContext preserves C:/extension on POSIX hosts
+  - newPotentialEntry preserves C:/extension on POSIX hosts
+Supporting Evidence:
+- Final targeted regressions: docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-targeted-regressions.md
+- Final coverage: docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-jest-coverage.md
