+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-targeted-regressions.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:42:01.8436215-04:00
+Command: Push-Location extensions/drm-copilot; npm run test:unit -- --runTestsByPath test/extension.test.ts test/repo-automation-service.test.ts -t 'helloPython|helloPowerShell|collectCommitContext|newPotentialEntry'; Pop-Location
+EXIT_CODE: 0
+Output Summary: Targeted final Windows-root POSIX regressions passed. Named passing scenarios: helloPython preserves C:/extension on POSIX hosts; helloPowerShell preserves C:/extension on POSIX hosts; collectCommitContext preserves C:/extension on POSIX hosts; newPotentialEntry preserves C:/extension on POSIX hosts. Jest reported 2 passed suites, 10 passed tests, 12 skipped tests, 22 total.
