+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/final-qa/final-tsc.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:41:53.1220066-04:00
+Command: Push-Location extensions/drm-copilot; npm run typecheck; Pop-Location
+EXIT_CODE: 0
+Output Summary: TypeScript type-checking passed with zero blocking diagnostics during the clean consecutive QA pass.
