+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/other/p1-t1.scope-lock.2026-04-05T20-08.md
@@ -1,0 +1,10 @@
+Timestamp: 2026-04-05T20-08
+Allowed Files:
+- extensions/drm-copilot/src/command-runtime.ts
+- extensions/drm-copilot/test/extension.test.ts
+- extensions/drm-copilot/test/repo-automation-service.test.ts
+- docs/features/active/2026-04-05-ci-failing-error-128/issue.md
+- docs/features/active/2026-04-05-ci-failing-error-128/plan.2026-04-05T19-46.md
+- docs/features/active/2026-04-05-ci-failing-error-128/evidence/**
+Out-of-Scope: any second production TypeScript file
+spec.md/user-story.md/research.md remain non-required minor-audit inputs
