+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/other/p1-t9.acceptance-mapping.2026-04-05T20-13.md
@@ -1,0 +1,16 @@
+Timestamp: 2026-04-05T20-13
+Requirements Source: docs/features/active/2026-04-05-ci-failing-error-128/issue.md
+Focused Regression Evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.*.md
+Planned Final QC Evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.*.md
+Only issue.md acceptance-criteria checkboxes may be changed from - [ ] to - [x] after supporting evidence exists; preserve the checkbox text exactly.
+
+Acceptance Criteria Mapping
+- [ ] Windows-style absolute extension roots such as `C:/extension` remain absolute when bundled script paths are resolved on POSIX hosts.
+  - Focused evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.*.md
+  - Final QC evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.*.md
+- [ ] Bundled command and repo-automation script invocations continue to target extension resources rather than workspace-relative copies.
+  - Focused evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.*.md
+  - Final QC evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.*.md
+- [ ] Regression coverage verifies the Windows-style mocked `fsPath` scenario and the extension Jest suite no longer fails with checkout-prefixed hybrid paths.
+  - Focused evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.*.md
+  - Final QC evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.*.md
