+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t1.format.2026-04-05T20-14.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20-14
+Command: npm --prefix extensions/drm-copilot run format
+EXIT_CODE: 0
+Output Summary: already formatted after final pass
