+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t2.lint.2026-04-05T20-14.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20-14
+Command: npm --prefix extensions/drm-copilot run lint
+EXIT_CODE: 0
+Output Summary: ESLint passed with no remaining blocking diagnostics.
