+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.2026-04-05T20-14.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20-14
+Command: npm --prefix extensions/drm-copilot run test:unit
+EXIT_CODE: 0
+Output Summary: All 10 extension Jest suites passed (10/10 suites, 140/140 tests), including the Windows-root bundled-path regression scenarios for helloPython, helloPowerShell, collectCommitContext, and newPotentialEntry.
