+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t6.clean-pass-summary.2026-04-05T20-14.md
@@ -1,0 +1,21 @@
+Timestamp: 2026-04-05T20-14
+Status: PASS
+Final QC Artifacts:
+- docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t1.format.2026-04-05T20-14.md
+- docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t2.lint.2026-04-05T20-14.md
+- docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t3.typecheck.2026-04-05T20-14.md
+- docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.2026-04-05T20-14.md
+
+Acceptance Criteria Results
+- PASS — Windows-style absolute extension roots such as `C:/extension` remain absolute when bundled script paths are resolved on POSIX hosts.
+  - Focused regression evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.2026-04-05T20-12.md
+  - Final QC evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.2026-04-05T20-14.md
+- PASS — Bundled command and repo-automation script invocations continue to target extension resources rather than workspace-relative copies.
+  - Focused regression evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.2026-04-05T20-12.md
+  - Final QC evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.2026-04-05T20-14.md
+- PASS — Regression coverage verifies the Windows-style mocked `fsPath` scenario and the extension Jest suite no longer fails with checkout-prefixed hybrid paths.
+  - Focused regression evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.2026-04-05T20-12.md
+  - Final QC evidence: docs/features/active/2026-04-05-ci-failing-error-128/evidence/qa-gates/p2-t4.test-unit.2026-04-05T20-14.md
+
+No Phase 2 command task was skipped.
+Non-Required Docs Confirmed: spec.md not required; user-story.md not required; research.md not required
