+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t6.red-jest.2026-04-05T20-11.md
@@ -1,0 +1,10 @@
+Timestamp: 2026-04-05T20-11
+Command: npm --prefix extensions/drm-copilot run test:unit -- --runTestsByPath test/extension.test.ts test/repo-automation-service.test.ts -t "helloPython|helloPowerShell|collectCommitContext|newPotentialEntry"
+EXIT_CODE: 1
+Output Summary: Focused red regression run failed as expected on the four Windows-root POSIX-host scenarios in extension and repo-automation service tests.
+Failure: helloPython preserves C:/extension on POSIX hosts
+Additional Failures:
+- helloPowerShell preserves C:/extension on POSIX hosts
+- collectCommitContext preserves C:/extension on POSIX hosts
+- newPotentialEntry preserves C:/extension on POSIX hosts
+First Diagnostic: Expected "C:/extension/resources/templates/hello_python.py" but received a checkout-prefixed hybrid path under "/Users/DanMoisan/repos/drm-copilot/extensions/drm-copilot/C:/extension/...".
