+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/regression-testing/p1-t8.green-jest.2026-04-05T20-12.md
@@ -1,0 +1,11 @@
+Timestamp: 2026-04-05T20-12
+Command: npm --prefix extensions/drm-copilot run test:unit -- --runTestsByPath test/extension.test.ts test/repo-automation-service.test.ts -t "helloPython|helloPowerShell|collectCommitContext|newPotentialEntry"
+EXIT_CODE: 0
+Output Summary: Focused green regression run passed for helloPython, helloPowerShell, collectCommitContext, and newPotentialEntry Windows-root scenarios.
+Passing Scenarios:
+- helloPython preserves C:/extension on POSIX hosts
+- helloPowerShell preserves C:/extension on POSIX hosts
+- collectCommitContext preserves C:/extension on POSIX hosts
+- newPotentialEntry preserves C:/extension on POSIX hosts
+Suite Totals: 2 passed, 2 total
+Test Totals: 26 passed, 25 skipped, 51 total
