+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/remediation-baseline/baseline-eslint.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:29:55.0902785-04:00
+Command: Push-Location extensions/drm-copilot; npm run lint; Pop-Location
+EXIT_CODE: 0
+Output Summary: ESLint baseline passed for the extension workspace with no blocking diagnostics.
