+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/remediation-baseline/baseline-jest-coverage.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:30:35.3315993-04:00
+Command: Push-Location extensions/drm-copilot; npx jest --coverage --coverageReporters=text-summary --runInBand; Pop-Location
+EXIT_CODE: 0
+Output Summary: Full extension Jest baseline passed with 10 passed suites and 140 passed tests. Coverage headline from the commanded run: lines 89.73%, statements 89.73%, functions 86.02%, branches 81.79%. Existing per-file coverage summary for the unchanged production fix reports src/command-runtime.ts at 93.46% lines and 92.30% functions.
