+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/remediation-baseline/baseline-line-counts.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:30:41.1203928-04:00
+Command: (Get-Content 'extensions/drm-copilot/test/extension.test.ts').Count; (Get-Content 'extensions/drm-copilot/test/repo-automation-service.test.ts').Count
+EXIT_CODE: 0
+Output Summary: Baseline touched-file line counts recorded before remediation. extensions/drm-copilot/test/extension.test.ts = 918 lines. extensions/drm-copilot/test/repo-automation-service.test.ts = 313 lines. The primary file-size violation is extension.test.ts at 918 lines.
