+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/remediation-baseline/baseline-prettier-check.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:29:41.2095031-04:00
+Command: Push-Location extensions/drm-copilot; npx prettier --check 'src/**/*.ts' 'test/**/*.ts' '*.json' '*.cjs'; Pop-Location
+EXIT_CODE: 0
+Output Summary: Prettier baseline check passed for extension TypeScript, test, JSON, and CJS files with "All matched files use Prettier code style!".
