+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/remediation-baseline/baseline-tsc.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-04-05T20:30:05.5536702-04:00
+Command: Push-Location extensions/drm-copilot; npm run typecheck; Pop-Location
+EXIT_CODE: 0
+Output Summary: TypeScript baseline type-check passed for the extension workspace with no blocking TSC diagnostics.
