+++ docs/features/active/2026-04-05-ci-failing-error-128/evidence/remediation-baseline/phase0-instructions-read.md
@@ -1,0 +1,19 @@
+Timestamp: 2026-04-05T20:29:03.6845955-04:00
+Policy Order:
+1. .github/copilot-instructions.md
+2. .github/instructions/general-code-change.instructions.md
+3. .github/instructions/general-unit-test.instructions.md
+4. .github/instructions/typescript-code-change.instructions.md
+5. .github/instructions/typescript-unit-test.instructions.md
+6. .github/instructions/typescript-suppressions.instructions.md
+Files Read:
+- c:\Users\DanMoisan\repos\drm-copilot\.github\copilot-instructions.md
+- c:\Users\DanMoisan\repos\drm-copilot\.github\instructions\general-code-change.instructions.md
+- c:\Users\DanMoisan\repos\drm-copilot\.github\instructions\general-unit-test.instructions.md
+- c:\Users\DanMoisan\repos\drm-copilot\.github\instructions\typescript-code-change.instructions.md
+- c:\Users\DanMoisan\repos\drm-copilot\.github\instructions\typescript-unit-test.instructions.md
+- c:\Users\DanMoisan\repos\drm-copilot\.github\instructions\typescript-suppressions.instructions.md
+Notes:
+- Repository tone and response policy confirmed before implementation work.
+- General code-change, unit-test, and TypeScript-specific policies were read before remediation.
+- The active feature folder is in minor-audit mode and issue.md remains the sole acceptance-criteria source.
