+++ docs/features/active/2026-04-05-ci-failing-error-128/feature-audit.2026-04-05T20-45.md
@@ -1,0 +1,65 @@
+# Feature Audit: 2026-04-05-ci-failing-error-128
+
+## Scope and Baseline
+
+- **Base Branch:** `development`
+- **Feature Folder:** `docs/features/active/2026-04-05-ci-failing-error-128/`
+- **Feature Folder Selection Rule:** Used the user-provided active feature folder and confirmed the same folder is referenced by the refreshed PR-context additional-context list.
+- **Evidence Sources:**
+  - Primary: `artifacts/pr_context.summary.txt` (refreshed during this review; commit-based)
+  - Secondary: `artifacts/pr_context.appendix.txt`
+  - Supplemental live evidence: `git status --short`, direct file inspection, and the live extension verification commands run during this review
+  - Feature evidence: `docs/features/active/2026-04-05-ci-failing-error-128/evidence/**`
+- **Work Mode:** `minor-audit`
+- **Authoritative AC Source:** `docs/features/active/2026-04-05-ci-failing-error-128/issue.md#acceptance-criteria`
+- **Non-required Docs Verification:** `spec.md` absent; `user-story.md` absent.
+- **Phase 0 Baseline Evidence Verification:** `evidence/baseline/` contains `phase0-instructions-read.md` and the required `p0-t2` through `p0-t8` baseline artifacts.
+- **Plan Checklist Reconciliation:** All checked plan items are backed by on-disk evidence files or direct file-content inspection of the touched source/test files.
+
+## Acceptance Criteria Inventory
+
+Authoritative acceptance criteria extracted only from the explicit `## Acceptance Criteria` section in `issue.md`:
+
+1. Windows-style absolute extension roots such as `C:/extension` remain absolute when bundled script paths are resolved on POSIX hosts.
+2. Bundled command and repo-automation script invocations continue to target extension resources rather than workspace-relative copies.
+3. Regression coverage verifies the Windows-style mocked `fsPath` scenario and the extension Jest suite no longer fails with checkout-prefixed hybrid paths.
+
+## Acceptance Criteria Evaluation
+
+| Criterion | Status | Evidence | Verification Command(s) | Notes |
+|---|---|---|---|---|
+| Windows-style absolute extension roots such as `C:/extension` remain absolute when bundled script paths are resolved on POSIX hosts. | PASS | `extensions/drm-copilot/src/command-runtime.ts`; `evidence/regression-testing/p1-t8.green-jest.2026-04-05T20-12.md`; live targeted regression run at 20:45 | `Push-Location extensions/drm-copilot; npm run test:unit -- --runTestsByPath test/extension.test.ts test/repo-automation-service.test.ts -t "helloPython|helloPowerShell|collectCommitContext|newPotentialEntry"; Pop-Location` | The drive-root guard remains in place and the named POSIX regression scenarios passed live. |
+| Bundled command and repo-automation script invocations continue to target extension resources rather than workspace-relative copies. | PASS | `extension.test.ts`, `repo-automation-service.test.ts`, `evidence/final-qa/final-targeted-regressions.md`, live targeted regression run at 20:45 | Same targeted Jest command as above | The preserved regression assertions continue to require `C:/extension/resources/templates/...` paths. |
+| Regression coverage verifies the Windows-style mocked `fsPath` scenario and the extension Jest suite no longer fails with checkout-prefixed hybrid paths. | PASS | `evidence/regression-testing/p1-t6.red-jest.2026-04-05T20-11.md`; `evidence/regression-testing/p1-t8.green-jest.2026-04-05T20-12.md`; `evidence/final-qa/final-jest-coverage.md`; live full unit-suite run at 20:45 (`11/11` suites, `140/140` tests) | `Push-Location extensions/drm-copilot; npm run test:unit; Pop-Location` | The feature has fail-before, pass-after, and fresh full-suite evidence. |
+
+## Additional Reduced-Audit Verification
+
+| Requirement | Status | Evidence | Notes |
+|---|---|---|---|
+| `spec.md` and `user-story.md` are absent | PASS | Feature-folder listing | Matches `minor-audit` requirements. |
+| Required Phase 0 baseline evidence exists | PASS | `evidence/baseline/phase0-instructions-read.md` and `p0-t2`–`p0-t8` artifacts | Baseline capture is complete. |
+| Plan checklist state is backed by evidence on disk | PASS | Baseline, regression, QA-gate, and remediation evidence folders plus direct source inspection | Checked checklist state is supportable from files on disk. |
+| Remediated touched test/helper files are each `<= 500` lines | PASS | `evidence/final-qa/final-scope-and-line-counts.md`; live line-count capture | All five touched test/helper files are below the cap. |
+| Code change remains within approved scope and remediation constraints | PASS | `issue.md`, `plan.2026-04-05T19-46.md`, `remediation-plan.2026-04-05T20-20.md`, live source inspection | Production scope remains the single approved fix; remediation broadens only companion test/helper layout as authorized. |
+
+## Summary
+
+**Overall feature readiness:** PASS
+
+The active small-path bug fix satisfies all three authoritative acceptance criteria from `issue.md`, the reduced-audit structural requirements are met, the remediation resolved the file-size policy blocker, and the live extension verification pass is green.
+
+**Top gaps preventing PASS:** None.
+
+**Recommended follow-up verification:** None required for the reduced audit. Commit the refreshed working-tree state and audit artifacts before PR preparation so PR-context-derived automation reflects the same content.
+
+## Acceptance Criteria Status
+
+- Source: `docs/features/active/2026-04-05-ci-failing-error-128/issue.md`
+- Total AC items: 3
+- Checked off (delivered): 3
+- Remaining (unchecked): 0
+- Items remaining: None
+
+## Acceptance Criteria Check-Off
+
+No checkbox edits were required during this re-review. All three authoritative acceptance-criteria items were already marked `[x]` in `issue.md`, and the current evidence supports those checked states.
